tests: create unittest_librbd
The existing ceph_test_librbd integration tests are now used
within unittest_librbd as unit tests run during 'make check'.
The unit tests use librados_test_stub to simulate RADOS.
